tests/wpt/tests/websockets/interfaces/WebSocket/send/002.html

<!doctype html>
<title>WebSockets: replacing send</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  ws.send = 5;
  assert_equals(ws.send, 5);
});
</script>
